backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic[email]>=2.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2

